pyproject.toml
[tool.poetry]
name = "reitinhaku"
version = "0.1.0"
description = ""
authors = ["Paavo Nätti <paavo.natti@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
pillow = "^9.4.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.2.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
